odb: shrink dbGCellGrid::GCellData fields uint -> uint8_t
Shrinkg 6_final.odb from 3.6G to 2.1G on gf12/bp_single.  The effect
on gzipped size is minimal as it effective compresses out this
inefficiency.

+++ src/gpl/src/routeBase.cpp
@@ -349,9 +349,9 @@ static float getUsageCapacityRatio(Tile* tile,
                                    odb::dbGCellGrid* gGrid,
                                    float ignoreEdgeRatio)
 {
-  unsigned int capH = 0, capV = 0, capU = 0;
-  unsigned int useH = 0, useV = 0, useU = 0;
-  unsigned int blockH = 0, blockV = 0, blockU = 0;
+  uint8_t capH = 0, capV = 0, capU = 0;
+  uint8_t useH = 0, useV = 0, useU = 0;
+  uint8_t blockH = 0, blockV = 0, blockU = 0;
   gGrid->getCapacity(layer, tile->x(), tile->y(), capH, capV, capU);
   gGrid->getUsage(layer, tile->x(), tile->y(), useH, useV, useU);
   gGrid->getBlockage(layer, tile->x(), tile->y(), blockH, blockV, blockU);
@@ -369,9 +369,9 @@ static float getUsageCapacityRatio(Tile* tile,
   // RePlAce/RC metric need 'blockage' because
   // if blockage ratio is larger than certain ratio,
   // need to skip.
-  unsigned int curCap = (isHorizontal) ? capH : capV;
-  unsigned int curUse = (isHorizontal) ? useH : useV;
-  unsigned int blockage = (isHorizontal) ? blockH : blockV;
+  uint8_t curCap = (isHorizontal) ? capH : capV;
+  uint8_t curUse = (isHorizontal) ? useH : useV;
+  uint8_t blockage = (isHorizontal) ? blockH : blockV;
 
   // escape tile ratio cals when capacity = 0
   if (curCap == 0) {
